refactored App.delegate code as well as socialFeed view to move get friends networking code to iPhone

+++ Nume WatchKit Extension/DictationController.swift
@@ -49,13 +49,15 @@ class DictationController: WKInterfaceController {
                     return
                 }
                 self.user!.userActivity! = input[0] as! String
+
                 
                 // Placing dictation text and number rating into NSUserDefaults
                 let appGroupID = "group.io.github.dhsu210.Nume"
                 let defaults = NSUserDefaults(suiteName: appGroupID)
                 defaults!.setValue(self.user!.userNumber, forKey: "userNumberKey")
                 defaults!.setValue(self.user!.userActivity, forKey: "userActivityKey")
-                self.user.userToken = defaults!.integerForKey("userTokenKey")
+                self.user.userToken = defaults!.integerForKey("userTokenKey") as Int
+
                 
                 // Pulls username from logged in FB account on iPhone
                 let defaultConnect = NSUserDefaults.standardUserDefaults()
@@ -70,7 +72,7 @@ class DictationController: WKInterfaceController {
                         println("Congratulations, \(responseName) successfully rated a \(self.user!.userNumber!) with '\(self.user!.userActivity!)'")
                     }
                 }
-
+                
                 self.pushControllerWithName("SocialFeed", context: self.user)
         }
